Introduce AnyInvocable
AnyInvocable a type-erased function wrapper - similar to `std::function`
- that wraps anything that fulfills the `std::invocable` concept. Just
like `std::invocable` - and unlink `std::function` - this wrapper does
not require the function object to be copyable and is itself not
copyable.

In other codebases this type is also often called `unique_function`. In
anticipation to replace this code by the upcoming Abseil implementation,
I named it `AnyInvocable` as this will probably be its name in Abseil.