Introduce AnyMovable
`AnyMovable` is a type-safe container for values of any type that
fulfills the std::movable concept. It's very similar to `std::any` with
the one exception that the stored values don't need to be copy
constructible. Hence `AnyMovable` is also not copy-constructible.

The API is also very similar to `std::any`. But the reference-based
overloads of `std::any_cast` are not implemented in
`orbit_base::any_movable_cast` because they rely on exceptions for error
handling. Instead you have to use the pointer overloads that return a
nullptr in case of a type mismatch.