Fix lifetime of continuations in ScheduleAfter
Continuations scheduled by MainThreadExecutor are destroyed after they
have been executed. If the MainThreadExecutor gets out of scope before
the continuation can be executed, it will be destroyed when the future
completes, and not when the MainThreadExecutor gets out-of-scope.

That's problematic behaviour if the continuation's destructor accesses
managed resources. That can happen for example when the continuation
captures a ScopedStatus that calls into `CaptureListener` in its
destructor.

To avoid that continuation are being destructed too late and potentially
in the wrong execution context (not on the main thread), this commit
changes ownership. The continuation will live in a
`std::list<std::AnyMovable>` in MainThreadExecutor. That ensures that
they don't live longer than MainThreadExecutor itself.