Introduce MainThreadExecutor::ScheduleAfterIfSuccess
`ScheduleAfterIfSuccess` is only available for
`Future<ErrorMessageOr<T>>`. It schedules a task that is executed when
the future completes, but only if the contained result is not an error
message. In case of an error the continuation will be bypassed and the
error message will be returned.

It's the semantic equivalent of an return-early-branch in synchronous
code.