Move implementation of Future into orbit_base_internal
In a subsequent commit, a specialization of Future<T> will be introduced.
This would lead to a lot of code duplication. The solution
is to move the current implementation of `orbit_base::Future` into
`orbit_base_internal::InternalFuture` and subclass that implementation.

Note that this is already the second subclass, since there is also
`orbit_base_internal::InternalFutureBase`. That's far from ideal but the
best solution to avoid code duplication.