Introduce ImmediateExecutor
ImmediateExecutor is a helper class to be used in conjunction with
`Future::Then`. It allows to chain a task to be executed whenever the
future completes. The task will be executed in the execution context
that sets the result of the promise.

This type of executor should only be used for tasks that don't depend on
any outside resources since it can't be guaranteed when and where this
task is executed.