Make --local connection faster
Orbit connects faster to a locally running OrbitService. Orbit does not
wait until the state of the local grpc connection is "ready". Instead
the connection is used immediately after setting it up. The local
connection is determined successful, when Orbit receives the first
response from OrbitService.

Adds a tool tip to the "Connecting..." text with more explanation.

Minor Refactoring in ProfilingTargetDialog.cpp