Fix GetCaptureTimeSpanUs for empty captures.
TimeGraph initializes the min and max capture time to
std::numeric_limits<uint64_t>::max and 0, leading to uint64_t max
as a result for GetCaptureTimeSpan for an empty capture and thus
not zero.

In fact there are checks on the caller-side, to not draw, e.g. the
scroll bar if the timespan is 0.

Test: Connect to an instance, see that the capture window does not
show a timeline.