Remove DataView::GetSelectedIndex
`GetSelectedIndex` actually returns the visible selected index. As
single selection can be considered as a special case of multi-selection,
`GetSelectedIndex` can be replace by `GetVisibleSelectedIndices` which
returns:
- an empty vector when `GetSelectedIndex` needs to return a nullopt;
- a single-value vector when `GetSelectedIndex` needs to return an index,

To be clear, `selected_indices_` represents the user's actual selection
in the corresponding view, either by mouse or arrow keys. If the user do
filtering or sorting without changing his selections,
`GetVisibleSelectedIndices` returns the intersection of
`selected_indices_` and visible indices. And we update the selected rows
in the correponding view according to the returned rows by
`GetVisibleSelectedIndices`.