Extend JoinFutures to arbitrary types
`JoinFutures` used to be only implemented for `Future<void>`. This new
change implements `JoinFutures` for `Future<T>`. It takes an
`absl::Span<const Future<T>>` as its argument and returns a
`Future<std::vector<T>>`. So far it's not possible to change the
container, that is returned. It's also not allocator-aware.