Introduce orbit_base::HandleErrorAndSetResultInPromise
It's a helper function for promises.

Assume you have `promise.SetResult(invocable(result.value()))`
while `result` is of type `ErrorMessageOr<T>`. This expression is
syntactically correct as long as `T != void` and `invocable` doesn't
return void.

This helper function implements that expression and does the right thing
if either of the two `void` occur.