Disable highlighting for FrameTrack
After adding frametrack, when clicking the corresponding function in the
live tab or the correponding function calls in the capture window, the
frametrack is also highlighted. This defeats the purpose of the
frametrack colors (i.e., indicating long frames) and so the selection
should not influence the frametrack.

With this change, we disable the highlighting for the frametrack. More
specifically:
- When clicking a function in live tab or a function call in the capture
window, the color of frametrack will not be affected;
- When clicking the frametrack, the frametrack will not be highlighted,
  but the other function calls with the same function id in the capture
  window and the function in the live tab will still be highlighted.