Erase TimeGraph->GetBatcher() dependence
Part of http://b/176086740.

We are erasing TimeGraph->GetBatcher() dependence by adding the
timegraph-batcher as a parameter of UpdatePrimitives functions.

Also I included a few renamings to differentiate it with the UI-batcher
used in Draw functions. We should use also the UI-batcher for the draw
functions, but this will come in another PR.