Track type cleanup
Remove unneeded track types and remove the logic depending on track
type in GetBackgroundColor.

With this, we could remove the need for track type completely, but it
still could be interesting to identify which data a track displays, so
I decided to leave it. We shouldn't use the type to emulate
polymorphism though, but only for informational purposes.