Add ScopeTree hierarchical data structure for profiling scopes
ScopeTree is a layer of abstraction above existing TextBox data. It
provides hierachical relationship between profiling scopes. The goal
is to be able to generate the scope tree with different streams of
scope data that can arrive out of order. The ScopeTree also maintains
and ordered map of nodes per depth.