Use protobuf Arenas in LockFreeBufferCaptureEventProducer
After various experimenting, I can say that this can save *a lot* of extra work
in `LockFreeBufferCaptureEventProducer::ForwarderThread`, specifically in terms
of heap allocations and deallocations.

The one downside is that it prevents moves from an already constructed protobuf,
which is the case for the Vulkan layer. I tested the two cases on Trata and
Infiltrator and didn't notice any difference: `ForwarderThread` was and still is
around 1 to 2 % CPU utilizaton.

Bug: http://b/179969217