Fix module reloading behaviour in OrbitApp
I accidentally changed the behaviour of module reloading when doing the
preset refactoring, by making `LoadModule` only load a module when it's
not yet loaded.

But the legacy overload of `LoadModules` (which is calling the new
`LoadModule`) expects the module to be reloaded when it is already loaded.

So this PR makes it obvious what's happening by introducing a bunch of
overloads with descriptive names.