Make TimeGraph optional in CaptureWindow. (#1864)
TimeGraph contains a lot of information/logic to draw capture
related information in the CaptureWindow. However, if we don't have
a capture yet, it will still be queried, e.g. for the capture duration.

This let to a couple of rendering bugs in the past. With this change
the time_graph of CaptureWindow is a unique_ptr (optional does not work
as TimeGraph is not easily movable) and is nullptr before a capture
was taken/loaded.

This also has the nice effect that the capture window is really
clean before taking a capture (no scrollbar, no time track, no
vertical green line).

Test: Try different actions (like clicks) on the empty capture window.
Load a capture and take a capture.