Add "ScopeTree" hierarchical data structure for profiling scopes
ScopeTree is a layer of abstraction above existing scope data. It
provides hierarchical relationship between profiling scopes. The goal
is to be able to generate the scope tree with different streams of
scope data that can arrive out of order. The ScopeTree also maintains
and ordered map of nodes per depth. This commit contains the data
structure and tests only, usage will come in another change.