Explicit construction of ErrorMessageOr in generics
`outcome::basic_result` disables its generic templated constructor if the
arguments could both construct the value and the error type to avoid any
kind of ambiguity. In this case the user is supposed to use the explicit
wrappers `outcome::success` and `outcome::failure`.

In generic code we should always use the explicit wrappers to avoid
problems with certain types that clash with `ErrorMessage`.

This commit adds the wrapper in Promise/Future code.