Expose setting to limit the local depth of markers per command buffer.
Adding a lot of Vulkan debug markers can have a significant performance
impact when capturing those.
On the layer, we already have an option to limit the depth of
markers (per command buffer). This change exposes this setting to the
capture settings dialog and transfer the specified value to the layer.

On the UI side, the setting is split into a boolean (checkbox) if
the depth should be limited at all and a text field (only enabled when
the checkbox is clicked and only allowing numeric strings in uint64 range).
Internally, we use uint64::max as to disable limitation of the depth
(as already done in the layer).

Note that the text fields also excepts empty strings (as otherwise
it is inconvinient to change the first digit), but resets those to
0.

Bug: http://b/180388662
Test: Try different settings in the capture options menu and take
 captures with Vulkan layer enabled.