Split module and symbole loading
The function `OrbitApp::LoadModule` used to retrieve a module file and
then load its symbols automatically. This PR changes that by renaming
`LoadModule` to `LoadModuleAndSymbols`.

Furthermore it introduces a new function called `OrbitApp::LoadModule`
which only retrieves the module and ensure that it is available in a
local file path. Maybe nothing needs to be done, because it's in a local
search path, or it needs to be copied from the remote instance and put
in the local cache directory. Either way, `LoadModule` will take care of
it and return the local path, if successful.

Test: Manual test
Bug: http://b/179241832