Initial structure for Jupyter notebooks and Python helper code
We want to add Jupyter notebook examples for prototyping and
deeper analysis of capture data. We add an initial structure here that
outlines most of the parts needed, including Python helper code
(in orbitutils), unit tests for Python helper code, a build script
(in Python so it works on Windows and Linux) that compiles the needed
protobuffers and runs all unit tests for verification, and a sample
notebook so users can verify their setup.