Introduce MainWindowInterface
This commit introduces an abstract base class which should simplify the
current callback situation. OrbitApp stores a huge number of callbacks in
std::function objects. Each needs to be set independently. There is no
compile-time mechanism to tell a user if they missed to set one of the
callbacks.

By introducing an abstract base class which can be implemented by
OrbitMainWindow, it gets easier to grasp the requirements. This base
class can be easily mocked in tests. Adding more pure virtual functions
will lead to compile-errors in the tests, meaning they can't be missed.

Furthermore this remove one indirection, compared to the std::function
solution, which means fewer heap allocations and faster code.

This commit also ports one callback over to the abstract base class.
This is needed, because without it it won't compiler since the member
variable in OrbitApp were not in use.