Implement simple case of "Go To Source"
With this commit users can display the source code of a certain function
when choosing the `Go To Source` action.

For now only the basic use case works, meaning the debug information is
not stripped and the file path in the debug information can be found on
the local machine without mapping to a different directory.