Fix crash on debug markers across multiple submissions.
If debug markers are across different submissions, and we captured
the "begin" marker in the Vulkan layer, but missed a tracepoint
for the "GpuJobs, we end up in a situation where we do have
the "begin" submission, but not the "begin" job.

We now handle this gracefully (similar to the case where we do not
have the "begin" submission), but using the begin capture time
as approximation.

Test: Use Vulkan layer on Infiltrator.