Fix the implementation of Vulkan extensions.
On the Vulkan layer, we want the game to be able to call the debug
marker extension methods (which are actually 2 different extensions).
In order for a game to know which extensions are available (by layer),
Vulkan provides two mechanisms.
1) EnumerateInstance/DeviceExtensionProperties function.
2) Declare them in the manifest, such that the loader can decide
on a queried extension, without loading the layer.

Prior this change, we did not exposed any extension in the manifest.
Also, the DebugUtils extension is an instance extension (and not a
device extension), so this adjusts the enumeration functions. As we
do not want to intercept the ResetQueryPoolEXT function, this extension
is removed from the enumeration function.
Further, the DebugMarkers extension depends on the DebugReport extension.
Thus, we also need to implement this.
Vulkan also dictates, that a layer implements all functions of an
extension, if it decides to implement that extension. So, this change
adds all the "unused" functions of the three extensions we are exposing.
For those functions, we simply call into the next layer (if possible).

TODO: Add tests for the VulkanController calling the empty functions.

Follow-up (different PR): Similar to a game, a layer needs to tell
the ICD/driver which extensions it is going to use (in vkCreate...).
Thus, we need to add the host query pool reset extension here.

Test: Run tests + profile on Infiltrator.