Add support for structured debug stores in SymbolHelper
This extends SymbolHelper's symbol search mechanism. In addition to
checking the paths in SymbolPaths.txt it will also check a structured
debug store. On Linux, this means searching in `/usr/lib/debug`. When
the Stadia SDK is installed this will also search in
`$GGP_SDK_PATH/sysroot/usr/lib/debug`.