Retry finalizing submits when querying for some timestamps fails
When finalizing submits it can happen that queries for timestamps can
fail when the timestamp is not yet ready to be queried. This aborted
with a CHECK for now, but needs to be worked around by re-enqueing the
submits where timestamp queries failed for later processing. This is
now fixed with this change.

Tested: Manually ran example application and capture multiple times.
Bug: http://179649352