Ensure to request required extension on vkCreateInstance/Device.
We use Vulkan's  host query reset extension. When creating a
device, this extension must be querried using the .ppEnabledExtensionNames
field of the create info. Before this, we were just lucky, as
most games also use that extension.
Similar the get physical device properties extension is required
transitively (via the host extension).

In order to be able to test
(i.e. exchange the vkEnumerateInstanceExtensionProperties), this change
introduces a "VulkanWrapper", which calls into the Vulkan loader
functions. Note, that we can't call the instance enumeration function
using the usual vkGetInstanceProcAddr, as this is a special function
by the Vulkan loader, where the calls directly into layers will fail,
as the loader collect all the extensions for all layers/ICDs.

Test: Run on hello_ggp_standalone & run tests.