LockFreeBufferCaptureEventProducer optimization
When dequeuing intermediate events, call the TranslateIntermediateEvents
virtual function once and pass in all dequeued events instead of calling
it as many time as there are dequeued events (up to 10K per iteration).

This also allows to generate capture events made of multiple
intermediate events, which will come in handy for a performance
investigation.