Clear StringManager at start of a capture.
With switching to non-hashes as string id's, we can't rely on
collision-resistence anymore. For every capture, string id's
would be used for different strings, thus resulting in showing
wrong strings in the UI.

This fixes the issue, by clearing StringManager in ClearCapture.
This was already be done for loaded captures.

Test: Take several captures, look at the GPU Track names.