VkFreeCommandBuffers also rollback the timer slots and remove the state.
A command buffer can actually also be "free"ed while it is in
recording or executable state. On free, we remove the command
buffer from our internal maps, but not from the state map.

This leads to a crash in OnCaptureStopped, as the device mapping
is gone, but the state remains.

This change prevents this, by also resetting the slots on Free
and erasing the state.

Bug: http://b/181217248
Test: Run tests or stop a capture at the end of infiltrator (when
the gpu is done)