Handle inline code in ElfFile::GetLineInfo
GetLineInfo used not to take inlining into account, meaning the
addresses from a particular function might point to source files all
over the place because parts have been inlined.

For the source code view we want samples to count for the code line in
the original function, not the inlined function.

So this change makes GetLineInfo take inlining into account and always
return the location of outmost inlining frame which is the function we
are interested in.