Integrate SourcePathsMapping into source code loading
This adds the source paths mapping step to the show source code process.

When the path found in debuginfo is not found locally, the source paths
mappings will be loaded and tried to apply. If no mapping succeeds the
user will be prompted to choose the file via the file open dialog.