Fix broken timers when command buffer begin timestamp is missing.
If we miss a command buffer's begin -- as we were not already
capturing -- we send "0" as special timestamp. However,
we use the first command buffer of a submission as a reference
to align the Gpu timestamps to our Cpu timeline. If this timestamp
was missing, we failed and resulted in timers with length of days.

This fixes this (for now) on the client site by discarding the
complete submission if the first command buffer has no begin
marker affected. Note that the "align" debug markers to the
begin of capture if their "begin" info is missing, remains working.

Follow up: We should re-consider sending incomplete submissions.

Test: Take a capture in the first actual frames of infiltrator.
Bug: http://b/181195585