Simplify use of locks in our Vulkan layer's SubmissionTracker class
We simplify the use of locks in the SubmissionTracker class to make it
easier to reason about correctness of the code and avoid sometimes
leaking query slots.

For example, in methods like MarkCommandBufferBegin, the mutex was
locked, then freed again, then it was checked if we were capturing,
then the mutex was locked again to record the timestamp. It was possible
that capturing was stopped in the middle and OnCaptureFinished called.
This lead to query slots that should have been cleaned up in
OnCaptureFinished to not get cleaned up at all.

We also tighten the check on whether we are capturing and directly tie
this check to the calls to OnCaptureFinished. The previous use of the
IsCapturing() method sometimes led to imcomplete command buffer states
that were neither cleaned up in OnCaptureFinished nor properly
completed.

With this change, all query slots are now properly cleaned up after the
call to OnCaptureFinished and after all remaining submits have been
completed.

Tested: Manually ran Vulkan layer with two sample applications.
Bug: http://b/181097789