Draw sample counts per line in source code
We want to include sample counts and percentages when showing source code
and disassembly view. This PR aims that. For that, we are creating a widget to
be in the right, where we show that information. Also, we are making the pop-up window wider.

Related to http://b/159809330 from Source Code feature.

To test: Simply run CodeViewerDemo.