Gracefully handle running out of query slots in our Vulkan layer
We simply CHECKed that we found a query slot so far, which sometimes
failed on some of our test applications. We now handle this case
gracefully by not aborting and continuing execution, however without
recording timestamps for command buffers and markers where we don't get
one of the slots (either begin or end or both).

Due to the logic in how we distribute slots, once we run out of slots
it usually happens that no further command buffer and debug marker
timeslices are created, as all slots are given to begin events. We free
them but freed ones are immediately given to begin slots again.

There is also a performance issue in NextReadyQuerySlot when running out
of slots as this function runs over all slots (currently 65536) every
time a slot is requested. This will be fixed in a follow-up change.

Tested: Manually ran layer on two sample applications with both small
and large numbers of total slots (100, 1000, 20000, 65535).
Bug: http://b/181097789