Make CodeReport::NumSamplesAtLine return an optional
This way we can differentiate between lines without any samples and
lines without any data.

The latter occurs in source code files. We always show the whole file
but there is only data available for a single function (might change in
the future). Hence we need a way to indicate to the user whether data is
available or not.