Optimize finding free timestamp query slots in our Vulkan layer
We optimize finding a free timestamp query slot. The previous code ran
over a vector of all possible slots until a free one (based on the
vector entry) was found. This was problematic when many, and in
particular all, slots were occupied already, leading to massive game
slowdown and a lot of CPU time spent in finding the next free slot.

This is mostly noticeable when running out of query slots. With a
maximum of 10'000 slots, the previous code had ~84% of exclusive
callstack samples in TimerQueryPool::NextReadyQuerySlot. For the new
code, this number went down to ~0.14%.

For validation purposes we keep the vector of slot states and check
allocated slots and reset slots against the expected state.

Tested: Manually ran Vulkan layer on two test applications. Profiled
Orbit layer / game performance using Orbit.
Bug: http://b/181639638