Don't crash on reused command buffers.
Command buffers can be reused without calling vkResetCommandBuffer
(followed by vkBegin/EndCommandBuffer). In this case, the timer
slots that are baked into the command buffer, need to be invalidated,
as we could have "Submit"->"Wait"->"Submit"->"Present" (note that
we are only querying for timeslots on present, and thus can not
uniquely identify the slots anymore).

This fixes this by introducing a two staged reset of slots.
We mark slots that we will not read anymore. Further we mark
slots that were part of a vkResetCommandBuffer.
The actual slot is only reset once marked with both.

Thus for the case of a re-use command buffer, we will never reset
the slot completely (at least unless vkResetCommandBuffer is called).

To not query resubmitted slots again, we keep the command buffer state
until reset/new begin, and added an optional timestamp for the submission.
This allows us to only persist (and therefore read) a slot if
it originates from this submission.

Test: Run Tests, Capture on Trata, Infiltrator and hello_ggp
Bug: http://b/181098020.