Fix source paths mapping
The source paths mapping had a problem when source and target path used
different file path separators.

This commit fixes that by using iteration over file path
segments for comparison. The previous approach compared strings instead
which obviously fails in that case.

There is also a test for mixed separators now. The additional test for
`InferMappingFromExample` was written to figure out if the problem lies
with that function. But that wasn't the case. I left the additional test
anyway.