Fix GGP and fuzz-testing build
The file `contrib/jupyter/build.py` is imported into `conanfile.py`
which is all good. But `contrib/jupyter/build.py` imports the package
`build` which is not available on all machines as it seems.

Since this package is not used in the file, I believe the proper fix is
to remove the `import` statement.