Avoid LoadSymbols in parallel
We used to check whether symbols are already loaded by checking
`module->is_loaded`, but that flag doesn't tell whether symbols are
currently being loaded. And this might lead to a race condition where
two competing `LoadSymbols` calls have been scheduled at the same time.

We avoid this by doing the same that we also do for module. We do
bookkeeping in a hash map about which module are currently being symbols
loaded for.