Fix PerfEventQueue::TopEvent,PopEvent returning different events
When the `PerfEvent` at the top of `priority_queue_of_events_not_ordered_by_fd_`
and the `PerfEvent` at the top of `heap_of_queues_of_events_ordered_by_fd_` have
the exact same timestamp, `TopEvent` was returning the former, while `PopEvent`
was popping and returning the latter. This was particularly problematic when
`PopEvent` is called based on the result of `TopEvent`.

Bug: http://b/181840879

Test: Follow instructions in the bug and verify it's fixed. Add unit test that
was previously failing and now succeeds.