Fix long running unit test.
AccessTraceesMemoryTest potentially accesses a lot of memory (6mb in one
run I checked). Since the access is done word by word with context
switches inbetween this takes too long for a unit test.

The test works just as well with limiting the memory access to 4kb.

If this should ever become a bottleneck there is a way of accessing
memory via a file system interface.

Bug: http://b/181880672
Test: Ran unit test locally.