Adjust frame track colors to be similar when durations are similar
When all frame times are fairly similar, we want all colors to be fairly
similar, as there is no need to point out small differences by using
a different color. We solve this by adjusting the maximum upwards a bit
when computing colors (and only then!) and when the minimum and
maximum are close to each other. Colors are still interpolated between
the real minimum and the adjusted maximum, so there will be still some
variation in color.

When the maximum is "large enough" compared with the minimum,
the behavior is not changed.

Tested: Manually tested frame track.
Bug: http://b/174224286