Enable parsing memory size unit
We collect the system-wide memory usage information from /proc/meminfo.
Each line of the file consists of a parameter name, followed by a colon,
the value of the parameter, and an option unit of measurement, for
example "MemTotal 1234567 kB".

Previously, when parsing this file, we assume the unit is fixed to be kB
and this assumption might be wrong. With this change, we will also parse
the memory size unit and convert the memory size value to be reported in
kilobytes.