Log and potentially warn about low clock resolution on the client
We had seen an issue where low clock resolution on our Windows client
led to highly inaccurate introspection timings (resolution was 1ms). We
now log and (in devmode) warn users when the estimated clock resolution
is low and hence timings may be inaccurate.

I tested this with the Windows clock we used previously and indeed
had to adjust some constants in the clock resolution estimation to even
get the right resolution.

Tested: Manually ran Orbit with -devmode to check if the warning
appears with the clock we used previously on Windows.
Bug: http://b/180995185