Add memory size unit check
According to the kernel code, the size unit in /proc/meminfo is fixed to
"kB" and "1 kB = 1 KiloBytes = 1024 Bytes". This is different from the
definition of KiloBytes introduced in http://en.wikipedia.org/wiki/Kilobyte.

With this change, instead of parsing the size unit, we check the size
unit string in the /proc/meminfo file. We also add comments to mention
that we report memory size in "kB" and consider "1 kB = 1024 Bytes" in the
code.