Simplify `CompleteSubmits` by using a priority queue.
Prior to this change, `CompleteSubmits` (which gets called
on `vkQueuePresetKHR`) started by calling `PullCompletedSubmissions`,
which estimated which submissions were finished (by inspecting the
last command buffer in there). However, it could still be
that those submissions were not actually finished, so
we implemented a push back mechanism. Also, we needed to sort
the result every time in `PullCompletedSubmissions`.

This change simplifies this logic, by maintaining a priority
queue, sorted (inverse) by the cpu timestamps.
In `CompleteSubmits`, we now iterate over that queue (starting
by the oldest submission), and immediately stop as soon as
the first timestamp is not ready. This way the code is less
complex and also we don't query results twice.

Test: Run tests & take captures on Infiltrator/Trata