Always compile Vulkan layer and use Conan.
This adds the Vulkan headers (1.1.121) to conan dependencies.
We now build the Vulkan layer on all Linux builds.

The "gen_lockfile.sh" script was run and and the changed hashes
of existing entries were reset.
In order to run the gen_lockfile.sh, conan needed to be upgraded.

To enable the builds for all Linux builds, the "with_vulkan_layer"
option was dropped.

For the validation layers header, the generated header was addeed
to third party and the cmake file was adjusted.

Test: Compile & run CI.
Bug: http://182137868.