Clean-up usages of Qt UI headers
Qt's autogenerated "ui_*.h" header files should not be included in
header files, as these can't be modeled in the buildsystem's dependency
model.

If a header file was read by the compiler before the included "ui_*.h"
was generated, that would lead to an error.

The solution: Only include the "ui_*.h" in cpp files. They should be
considered implementation details anyway.