Add orbit_code_viewer::OwningDialog
OwningDialog is a version of Dialog that owns all its needed resources.
So far this only impacts the CodeReport object, since the syntax
highlighter needs to be owned anyway. But more might come in the future.

Furthermore the free function `OpenAndDeleteOnClose` is provided. It is
meant as an alternative to `QDialog::exec`.