Automatically adjust unit for "System Memory Usage"& "Production Limit"
With this change, the size units (e.g., GB or MB) for the "System Memory Total"
and "Production Limits" can to be automatically adjusted according to the value.
To do this, we change the `GraphTrack::warning_threshold_` and
`GraphTrack::value_upper_bound_` to hold the pretty name (e.g., "Memory Total:
8.00GB") and the raw value (e.g., 8399608 KB), rather than the label (e.g.,
"Memory Total") and the raw value.