Make ReadProcessMemory handle partial reads
ReadProcessMemory so far was not handling partial reads and left that to
the user. This is a behaviour change, but since the only user of that
endpoint (Disassembly view) was not handling that on the call side
either, I thought this shoule be fixed.

So far I haven't seen any partial reads though.