Use GetDeclarationLocationOfFunction in ShowSourceCode
`OrbitApp::ShowSourceCode` used to determine the jump location in a
source code file by getting LineInfo for the first instruction of the
function. This LineInfo does not necessarily need to be associated with
the first C++-statement.

We actually wanna jump to the first line of the function's declaration,
not to the first statement. So this commit introduces
`GetDeclarationLocationOfFunction` to determine the correct first source
code line.