Allocate memory suitable for trampolines.
Previously the allocation in the tracee happened at arbitrary memory
positions. For the trampolines we need chunks of memory in a 32 bit
offset range of the executable code segments of the trace.

This change contains:

a) Changes to facilitate allocation of memory at a specified position
(compare AllocateInTracee.h) and follow up changes to this.

b) The logic to determine the position of the memory chunk for the
trampolines (compare Trampoline.h) and the code doing the actual
allocation.

c) Fix for a glitch in FreeInTracee: Previously it was possible that
FreeInTracee used the memory segment it was meant to unmap as an
intermediate working area (which causes a crash).