Update CaptureWindow position after filtering
We had the issue that after filtering sometimes the capture windows
doesn't show anything because as we are not showing some tracks anymore,
the old position doesn't have any track. In this PR we ask for an update
after updating tracks position, so the view y-position is clamped for the
new top-y and bottom-y coordinates.

Something to consider:
 - I don't like TrackManager asking CaptureWindows to update position.
 We should have a better way to do this (CaptureWindows updating its
 position after drawing, but as it wasn't that simple) I left as it is
 now.

 - I overloaded the method UpdateWorldTopLeftY because I feel that it is
 nicer (can be convinced about the opposite)

 - This solution came from the idea in
 https://github.com/google/orbit/pull/2089 but without splitting
 UpdateTrackPosition and UpdateTrackPrimitives

Solve http://b/182978978.
Test: Load a capture, move the screen, filter.