Fix orders of production limit line & white label
The previous code puts the production limit line and the white label in
the same layer. As we draw primitives in the same layer in the order of
"Box --> Line --> Triangle", the red line of production limit appears
over the white label.

With this change, we fix the wrong ordering by putting the while label
in a different layer with the production limit.