Replace Set method used in encoding memory events
With this change, we replace the `RepeatedField<Element>::Set(...)`
method we used when encoding memory events into `TimerInfo` registers
by the `RepeatedField<Element>::Add(...)` method.

The `Set` method will not update the `current_size_` of the repeated
field. If `current_size_ == 0`, the repeated filed will not be copied
when copying the protobuf (either using the = operator or using the
`CopyFrom` method). This results in the loss of memory information in
the registers if we want to save the memory TimerInfo into
`GraphTrack::timers_`.