Fix y-coordinate of GPU timers with command buffers.
On a GPU track, we have a small gap (2px) between the different depths.
When we have command buffer timers, the depth is multiplied by 2.
However, as there is no gap between the hw execution timer and the
command buffer timers, the number of gaps must remain the same in both cases.

Test: Take a capture with the Vulkan layer and expand the track.
Bug: http://b/184720175